+++ android/app/src/main/java/de/dexpro/headsup/MainActivity.java
@@ -1,0 +1,5 @@
+package de.dexpro.headsup;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}